+++ lib/elixir/lib/integer.ex
@@ -16,4 +16,28 @@ defmodule Integer do
   defmacro even?(n) do
     quote do: rem(unquote(n), 2) == 0
   end
+
+  @doc """
+  Converts a binary to an integer. If successful, returns a
+  tuple of the form `{ integer, remainder_of_binary }`.
+  Otherwise `:error`.
+
+  ## Examples
+
+      iex> Integer.parse("34")
+      {34,""}
+      iex> Integer.parse("34.5")
+      {34,".5"}
+      iex> Integer.parse("three")
+      :error
+
+  """
+  @spec parse(binary) :: { integer, binary } | :error
+  def parse(binary) do
+    { result, remainder } = :string.to_integer(:binary.bin_to_list(binary))
+    case result do
+      :error -> :error
+      _ -> { result, :binary.list_to_bin(remainder) }
+    end
+  end
 end